% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adoption_data}
\alias{adoption_data}
\title{Simulated Dataset of Time and Adoptionstatus}
\format{
A data frame with 50 rows and 2 variables:
\describe{
  \item{time_index}{(integer) Point in Time of Measure}
  \item{adoption}{(integer) Dichotomous integer variable indicating wether or not an individual adopted a technology or innovation}
}
}
\usage{
adoption_data
}
\description{
This simulated dataset contains information about the point in time and the status of adoption of an individual
}
\keyword{datasets}
