% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-bass-model.R
\name{bayesian_bass_model}
\alias{bayesian_bass_model}
\title{Specification of Bayesian Bass Model}
\usage{
bayesian_bass_model(p_mean = 1e-06, p_sdev = 0.5, q_mean = 1e-06,
  q_sdev = 0.5)
}
\value{
A character string containing the Bayesian Bass Diffusion Model specifications
}
\description{
Function returns a string with model specifications that can be processed by rjags
}
