% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.bayesian_bass}
\alias{coef.bayesian_bass}
\title{Extract Model Coefficients}
\usage{
\method{coef}{bayesian_bass}(x)
}
\arguments{
\item{x}{An bayesian_bass object}
}
\value{
A numeric vector containing the coeficients where the first element is p and the second is q
}
\description{
A function to extract the Bass model coefficients
}
\examples{
\dontrun{
coef(bayesian_bass)
}
}
