% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-range.R
\name{compute_range}
\alias{compute_range}
\title{Compute Range of a variable}
\usage{
compute_range(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector of length 2 where the first element is the lower boundry and the second element is the upper boundry
}
\description{
Computes a reasonable range by adding the standard deviation of a variable to the minimum and maximum value of the variable
}
\examples{
compute_range(x = rnorm(100, sd = 4))
x <- rnorm(100)
y <- rnorm(100, sd = 40)
plot(x, y, xlim = compute_range(x), ylim = compute_range(y))
}
