% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bayesian-bass-model.R
\name{plot.bayesian_bass}
\alias{plot.bayesian_bass}
\title{Plot Bayesian Bass Model}
\usage{
\method{plot}{bayesian_bass}(x, se = F, ...)
}
\arguments{
\item{x}{An object of class bayesian_bass}

\item{se}{Boolian, default FALSE, if TRUE plots 95\% bernoulli confidence intervals of actual observed adoption rates}

\item{...}{Arguments to be passed to methods, such as graphical parameters}
}
\value{
A ggplot object
}
\description{
Plots the diffusion curve and the actual observed adoption rates on a timeline plot
}
\examples{
data(adoption_data)
model <- bayesian_bass_model()
fit <- bayesian_bass(data = adoption_data, var = "adoption", model = model)
plot(fit)

}
