% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bayesian-bass-model.R
\name{plot.bayesian_bass_diffusion}
\alias{plot.bayesian_bass_diffusion}
\title{Plot Diffusion Curve}
\usage{
\method{plot}{bayesian_bass_diffusion}(obj, hdi = TRUE)
}
\arguments{
\item{obj}{A Bayesian Bass model}
}
\description{
Plot Diffusion Curve
}
\examples{
data(adoption_data)
model <- bayesian_bass_model()
fit <- bayesian_bass(data = adoption_data, var = "adoption", model = model)
fit_diffusion <- predict_diffusion(fit)
plot(fit_diffusion)

}
