% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-bass-model.R
\name{posterior}
\alias{posterior}
\title{Calcualte the posterior distribution of a Bass Diffusion model prediction}
\usage{
posterior(x, time_index = 1)
}
\arguments{
\item{time_index}{}
}
\description{
This function returns a prediction based on the MCMC sampled rjags chains.
The point in time of prediction corresponds to the given time_index
}
\examples{
data(adoption_data)
model <- bayesian_bass_model()
fit <- bayesian_bass(data = adoption_data, var = "adoption", model = model)
posterior(fit, time_index = 1)
}
