% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-bass-model.R
\name{predict_adoption}
\alias{predict_adoption}
\title{Predicts the adoption rate}
\usage{
predict_adoption(obj, time)
}
\arguments{
\item{obj}{An object of class \code{bayesian_bass}}

\item{time}{An integer value specifing the point in time}
}
\description{
The adoption rate is defined as the proportion of adopters of a specific point of time
}
\examples{
data(adoption_data)
model <- bayesian_bass_model()
fit <- bayesian_bass(data = adoption_data, var = "adoption", model = model)
predict_adoption(fit, time = 1)
}
