% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-bass-model.R
\name{predict_bass}
\alias{predict_bass}
\title{Internal Function used by different predict functions}
\usage{
predict_bass(p, q, time)
}
\value{

}
\description{
Internal Function used by different predict functions
}
