% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bayesian_bass.R
\name{print.bayesian_bass}
\alias{print.bayesian_bass}
\title{Print bayesian bass model}
\usage{
\method{print}{bayesian_bass}(x, ...)
}
\arguments{
\item{x}{bayesian model}

\item{...}{}
}
\value{

}
\description{
Print bayesian bass model
}
