% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-bass-model.R
\name{predict.bayesian_bass}
\alias{predict.bayesian_bass}
\title{Bass Model Predictions}
\usage{
\method{predict}{bayesian_bass}(object)
}
\arguments{
\item{object}{An \code{bayesian_bass} object}
}
\value{
A tibble with two columns, time and prediction
}
\description{
Bass Model Predictions
}
\examples{
\dontrun{predict(bass_model)}
}
